backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.9.0
python-jose[cryptography]==3.5.0
httpx>=0.27.0
pydantic>=2.10.0
pydantic-settings==2.14.0
openai>=1.50.0
stripe>=11.0.0
python-multipart>=0.0.12
slowapi==0.1.9
tenacity>=8.5.0
python-dateutil>=2.9.0
reportlab>=4.2.0
resend>=2.0.0
apscheduler>=3.10.0
twilio>=9.0.0
openpyxl>=3.1.0
